app/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
PyPDF2
openai
python-dotenv

